// hr-frontend/src/lib/api/generated/model/organizationSummary.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { OrganizationCounts } from './organizationCounts';

export interface OrganizationSummary {
  id?: number;
  code?: string;
  nameEn?: string;
  nameAr?: string;
  organizationTypeId?: number;
  type?: string;
  typeCode?: string;
  status?: string;
  establishedDate?: string;
  logoText?: string;
  summary?: OrganizationCounts;
}